+++ zootopia/src/main/resources/mybatis/member-mapper.xml
@@ -81,7 +81,6 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 SELECT member_id FROM cert WHERE secret=#{secret}
 </select>
 
-
 </mapper>
